feat: Extract Wand button and console IO to separate classes
